fix(predictioneval): carry all twelve int64 wire fields losslessly
The ordered-rules model carries twelve exported int64 JSON fields that are
quantities and positions rather than bit patterns: supplied values and
availability positions, the declared interval endpoints, candidate,
intervention and cutoff positions, the four positions a result re-exports,
and the exact pool sum. They travelled as JSON numbers, and a JSON number
is read by an IEEE-754 consumer as a float64, which cannot hold every
int64.

Measured before this change: a source relayed through such a consumer had
supplied points of 2^53+1 arrive as 2^53, the exact int64 pool sum shifted
with them, and the recorded share moved from 0x3feffffffffffffe to
0x3ff0000000000000 — exactly 1.0, reporting an outcome as holding the
whole pool when it holds all of it but one point. Same status, same chosen
outcome, silently different evidence.

Introduce OrderedRulesInt64: an int64 in memory, a canonical base-ten
string on the wire. One shared type rather than twelve hand-written
serializers, and deliberately NOT OrderedRulesHex64 — hexadecimal is for
values whose identity is their bits, decimal for values whose identity is
their magnitude. Canonicality is checked by inversion rather than by a
list of rejected spellings: UnmarshalText parses and then requires
FormatInt(parsed, 10) to equal the input byte for byte, which makes the
decoder the encoder's inverse by construction. No numeric compatibility
shim and no float64 fallback; encoding/json will not hand a number to a
TextUnmarshaler at all. Negative values are carried — the int64 domain is
not narrowed to fit a consumer.

All arithmetic stays on int64 and every conversion is exact in both
directions, so the mechanism semantics are unchanged and no digest moved.
No version constant is bumped: this is transport representation with no
semantic content, on a Draft PR with no production runtime caller.

The relay proof carries its own control — the same helper is first shown
still destroying a bare int64, so a relay that had become lossless could
not make the assertions pass vacuously. After the repair the points, the
pool sum, the share, all four digests, the stream's selection-digest
verification and two positions one apart all survive it.

Coverage is two halves of different kinds. A reflective walk of the
exported type graph requires every JSON-tagged 64-bit field to be one of
the two wire types and finds sixteen; a hand-written list names the twelve
and walks each through MinInt64, -2^53-1, -2^53, -2^53+1, -1, 0, 1,
2^53-1, 2^53, 2^53+1 and MaxInt64 in its real carrier. The fifteen
platform-width int fields are measured rather than waved through: the one
a caller supplies is refused above the candidate ceiling and is not
re-exported by the refusal.

Thirteen mutants: twelve killed at the exact seam, one classified as
equivalent — changing the parse base from 10 to 0 kills nothing, because
canonicality already refuses every spelling the wider base would admit.
That is recorded in the code, so a reader does not credit the base for
strictness it does not provide.

No property-based-testing library was added; go.mod stays untouched and
the domain coverage is deterministic instead.

+++ internal/predictioneval/ordered_rules_digest.go
@@ -70,9 +70,9 @@ func digestSupplied(h hash.Hash, v SuppliedInt64) {
 	// behind a MISSING presence would make "absent" and "absent, with a
 	// leftover zero" produce different digests for the same evidence.
 	if v.Presence == SuppliedKnown {
-		digestInt(h, v.Value)
+		digestInt(h, int64(v.Value))
 		digestPart(h, v.Provenance)
-		digestInt(h, v.AvailableAtPosition)
+		digestInt(h, int64(v.AvailableAtPosition))
 		// The DECLARATION as well as the position. Inside a projected stream
 		// the flag is always set, since the projection refuses a KNOWN value
 		// without it — which is exactly why leaving it unhashed looked like
@@ -92,7 +92,7 @@ func digestSupplied(h hash.Hash, v SuppliedInt64) {
 
 func digestCandidate(h hash.Hash, c *OrderedRulesCandidate) {
 	digestPart(h, c.Identity)
-	digestInt(h, c.Position)
+	digestInt(h, int64(c.Position))
 	digestBool(h, c.HasPosition)
 	digestPart(h, string(c.SourceKind))
 	digestPart(h, string(c.EpisodeMembership))
@@ -126,8 +126,8 @@ func orderedRulesStreamDigest(s OrderedRulesStream) string {
 	digestPart(h, s.Scope.SourceContractVersion)
 	digestPart(h, string(s.Scope.Coverage))
 	digestPart(h, s.Scope.CoverageDetail)
-	digestInt(h, s.Scope.IntervalFromPosition)
-	digestInt(h, s.Scope.IntervalToPosition)
+	digestInt(h, int64(s.Scope.IntervalFromPosition))
+	digestInt(h, int64(s.Scope.IntervalToPosition))
 	digestBool(h, s.Scope.HasInterval)
 
 	digestPart(h, s.Admission.ManifestID)
@@ -140,7 +140,7 @@ func orderedRulesStreamDigest(s OrderedRulesStream) string {
 	}
 
 	digestBool(h, s.Cutoff.Established)
-	digestInt(h, s.Cutoff.Position)
+	digestInt(h, int64(s.Cutoff.Position))
 	digestPart(h, string(s.Cutoff.Kind))
 	digestPart(h, s.Cutoff.Identity)
 	digestPart(h, string(s.Cutoff.Basis))
@@ -276,7 +276,7 @@ func orderedRulesConsumedDigest(s OrderedRulesStream, configDigest string, d Sup
 	// validateScope refuses a stream that omits it, so no ADMITTED prefix can
 	// differ by it; hashing it here would separate two refusals from each
 	// other and nothing else.
-	digestInt(h, s.Scope.IntervalFromPosition)
+	digestInt(h, int64(s.Scope.IntervalFromPosition))
 	// And the END of the declared window, but ONLY when the traversal stopped
 	// because it ran out of stream.
 	//
@@ -310,7 +310,7 @@ func orderedRulesConsumedDigest(s OrderedRulesStream, configDigest string, d Sup
 	terminal := exhausted && !s.Cutoff.Established
 	digestBool(h, terminal)
 	if terminal {
-		digestInt(h, s.Scope.IntervalToPosition)
+		digestInt(h, int64(s.Scope.IntervalToPosition))
 	}
 	digestPart(h, s.Admission.ManifestID)
 	digestPart(h, string(s.Admission.ViewKind))
@@ -326,7 +326,7 @@ func orderedRulesConsumedDigest(s OrderedRulesStream, configDigest string, d Sup
 	// Its DroppedAtOrAfter count is not: that is a fact about candidates the
 	// traversal never reached.
 	digestBool(h, s.Cutoff.Established)
-	digestInt(h, s.Cutoff.Position)
+	digestInt(h, int64(s.Cutoff.Position))
 	digestPart(h, string(s.Cutoff.Kind))
 	digestPart(h, s.Cutoff.Identity)
 	digestPart(h, string(s.Cutoff.Basis))
